refactor: move SpecTypes out of xy_chart
