+++ src/chart_types/xy_chart/crosshair/crosshair_utils.ordinal_snap.test.ts
@@ -1,10 +1,11 @@
 import { computeXScale } from '../utils/scales';
-import { BasicSeriesSpec, SpecTypes, SeriesTypes } from '../utils/specs';
+import { BasicSeriesSpec, SeriesTypes } from '../utils/specs';
 import { getGroupId, getSpecId } from '../../../utils/ids';
 import { ScaleType } from '../../../utils/scales/scales';
 import { getSnapPosition } from './crosshair_utils';
 import { computeSeriesDomains } from '../state/utils';
 import { ChartTypes } from '../..';
+import { SpecTypes } from '../../../specs/settings';
 
 describe('Crosshair utils ordinal scales', () => {
   const barSeries1SpecId = getSpecId('barSeries1');
