+++ src/chart_types/xy_chart/domains/y_domain.test.ts
@@ -1,6 +1,6 @@
 import { ScaleType } from '../../../utils/scales/scales';
 import { RawDataSeries } from '../utils/series';
-import { BasicSeriesSpec, DomainRange, SpecTypes, SeriesTypes } from '../utils/specs';
+import { BasicSeriesSpec, DomainRange, SeriesTypes } from '../utils/specs';
 import { BARCHART_1Y0G } from '../../../utils/data_samples/test_dataset';
 import {
   coerceYScaleTypes,
@@ -11,6 +11,7 @@ import {
 } from './y_domain';
 import { GroupId } from '../../../utils/ids';
 import { ChartTypes } from '../..';
+import { SpecTypes } from '../../../specs/settings';
 
 describe('Y Domain', () => {
   test('Should merge Y domain', () => {
