+++ src/chart_types/xy_chart/legend/legend.test.ts
@@ -2,8 +2,9 @@ import { getAxisId, getGroupId, getSpecId } from '../../../utils/ids';
 import { ScaleType } from '../../../utils/scales/scales';
 import { computeLegend } from './legend';
 import { SeriesCollectionValue, getSeriesLabel } from '../utils/series';
-import { AxisSpec, BasicSeriesSpec, Position, SpecTypes, SeriesTypes } from '../utils/specs';
+import { AxisSpec, BasicSeriesSpec, Position, SeriesTypes } from '../utils/specs';
 import { ChartTypes } from '../..';
+import { SpecTypes } from '../../../specs/settings';
 
 const nullDisplayValue = {
   formatted: {
