+++ src/chart_types/xy_chart/rendering/rendering.bars.test.ts
@@ -3,10 +3,11 @@ import { identity } from '../../../utils/commons';
 import { ScaleType } from '../../../utils/scales/scales';
 import { renderBars } from './rendering';
 import { computeXScale, computeYScales } from '../utils/scales';
-import { BarSeriesSpec, DomainRange, SpecTypes, SeriesTypes } from '../utils/specs';
+import { BarSeriesSpec, DomainRange, SeriesTypes } from '../utils/specs';
 import { LIGHT_THEME } from '../../../utils/themes/light_theme';
 import { GroupId } from '../../../utils/ids';
 import { ChartTypes } from '../..';
+import { SpecTypes } from '../../../specs/settings';
 
 const SPEC_ID = 'spec_1';
 const GROUP_ID = 'group_1';
