+++ src/chart_types/xy_chart/specs/axis.tsx
@@ -1,6 +1,7 @@
-import { AxisSpec, Position, DEFAULT_GLOBAL_ID, SpecTypes } from '../utils/specs';
+import { AxisSpec, Position, DEFAULT_GLOBAL_ID } from '../utils/specs';
 import { ChartTypes } from '../../../chart_types';
 import { specComponentFactory, getConnect } from '../../../state/spec_factory';
+import { SpecTypes } from '../../../specs/settings';
 
 const defaultProps = {
   chartType: ChartTypes.XYAxis,
