+++ src/chart_types/xy_chart/specs/bar_series.tsx
@@ -1,7 +1,8 @@
-import { BarSeriesSpec, DEFAULT_GLOBAL_ID, SpecTypes, SeriesTypes } from '../utils/specs';
+import { BarSeriesSpec, DEFAULT_GLOBAL_ID, SeriesTypes } from '../utils/specs';
 import { ScaleType } from '../../../utils/scales/scales';
 import { ChartTypes } from '../../../chart_types';
 import { specComponentFactory, getConnect } from '../../../state/spec_factory';
+import { SpecTypes } from '../../../specs/settings';
 
 const defaultProps = {
   chartType: ChartTypes.XYAxis,
