+++ src/chart_types/xy_chart/state/chart_state.interactions.test.ts
@@ -1,9 +1,9 @@
 import { createStore, Store } from 'redux';
-import { BarSeriesSpec, BasicSeriesSpec, AxisSpec, Position, SpecTypes, SeriesTypes } from '../utils/specs';
+import { BarSeriesSpec, BasicSeriesSpec, AxisSpec, Position, SeriesTypes } from '../utils/specs';
 import { TooltipType } from '../utils/interactions';
 import { ScaleType } from '../../../utils/scales/scales';
 import { chartStoreReducer, GlobalChartState } from '../../../state/chart_state';
-import { SettingsSpec, DEFAULT_SETTINGS_SPEC } from '../../../specs';
+import { SettingsSpec, DEFAULT_SETTINGS_SPEC, SpecTypes } from '../../../specs';
 import { computeSeriesGeometriesSelector } from './selectors/compute_series_geometries';
 import { getProjectedPointerPositionSelector } from './selectors/get_projected_pointer_position';
 import {
