+++ src/chart_types/xy_chart/state/selectors/get_specs.ts
@@ -1,9 +1,10 @@
 import createCachedSelector from 're-reselect';
 import { GlobalChartState } from '../../../../state/chart_state';
 import { getSpecsFromStore } from '../../../../state/utils';
-import { AxisSpec, BasicSeriesSpec, AnnotationSpec, SpecTypes } from '../../utils/specs';
+import { AxisSpec, BasicSeriesSpec, AnnotationSpec } from '../../utils/specs';
 import { getChartIdSelector } from '../../../../state/selectors/get_chart_id';
 import { ChartTypes } from '../../..';
+import { SpecTypes } from '../../../../specs/settings';
 
 const getSpecs = (state: GlobalChartState) => state.specs;
