+++ src/specs/index.ts
@@ -8,6 +8,7 @@ export interface Spec {
   /** The type of spec, can be series, axis, annotation, settings etc*/
   specType: string;
 }
+
 export * from './settings';
 
 export * from '../chart_types/specs';
