+++ src/state/selectors/get_settings_specs.ts
@@ -2,9 +2,8 @@ import createCachedSelector from 're-reselect';
 import { GlobalChartState } from '../chart_state';
 import { ChartTypes } from '../../chart_types';
 import { getSpecsFromStore } from '../utils';
-import { SettingsSpec } from '../../specs/settings';
+import { SettingsSpec, SpecTypes } from '../../specs/settings';
 import { getChartIdSelector } from './get_chart_id';
-import { SpecTypes } from '../../chart_types/xy_chart/utils/specs';
 
 const getSpecs = (state: GlobalChartState) => state.specs;
